requirements.txt
astroid==2.5.3
azure-functions==1.7.0
certifi==2020.12.5
chardet==4.0.0
decorator==5.0.7
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
py==1.10.0
pylint==2.7.4
requests==2.25.1
retry==0.9.2
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
